requirements.txt
affine==2.3.1
altair==4.2.0
altgraph==0.17.2
appdirs==1.4.4
attrs==22.1.0
autopep8==1.7.0
bcdata==0.7.4
beautifulsoup4==4.11.1
blinker==1.5
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.5
commonmark==0.9.1
cryptography==37.0.4
cssselect==1.1.0
cssselect2==0.7.0
cycler==0.11.0
decorator==5.1.1
easygui==0.98.3
entrypoints==0.4
et-xmlfile==1.1.0
fake-useragent==0.1.11
fiona==1.8.21
flake8==5.0.4
Flask==2.2.2
fonttools==4.34.4
future==0.18.2
fuzzywuzzy==0.18.0
GDAL @ file:///Z:/PYTHON_SETUP/wheels/GDAL-3.4.3-cp39-cp39-win_amd64.whl
GeoAlchemy2==0.12.5
geopandas==0.11.1
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.3
halo==0.0.31
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
kiwisolver==1.4.4
laspy==2.4.1
lazrs==0.5.0
log-symbols==0.0.14
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
mccabe==0.7.0
munch==2.5.0
mypy-extensions==0.4.3
numpy==1.23.1
opencv-python==4.6.0.66
openpyxl==3.0.10
oracledb==1.0.3
OWSLib==0.27.2
packaging==21.3
pandas==1.4.3
parse==1.19.0
pathspec==0.10.1
pefile==2022.5.30
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
progressbar2==4.0.0
protobuf==3.20.1
psutil==5.9.2
psycopg2==2.9.3
py==1.11.0
pyarrow==9.0.0
pycodestyle==2.9.1
pycparser==2.21
pydeck==0.8.0b3
pyee==8.2.2
pyflakes==2.5.0
Pygments==2.13.0
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
Pympler==1.0.1
pyparsing==3.0.9
PyPDF2==2.9.0
pyppeteer==1.0.2
pyproj==3.3.1
pyquery==1.4.3
pyrsistent==0.18.1
pytest==7.1.3
python-dateutil==2.8.2
python-utils==3.3.3
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
PyYAML==6.0
rasterio==1.3.0
reportlab==3.6.11
requests==2.28.1
requests-html==0.10.0
rich==12.5.1
Rtree==1.0.0
scipy==1.9.1
semver==2.13.0
Shapely==1.8.2
six==1.16.0
smmap==5.0.0
snuggs==1.4.7
soupsieve==2.3.2.post1
spinners==0.0.24
SQLAlchemy==1.4.41
streamlit==1.13.0
svg2rlg==0.3
svglib==1.4.1
termcolor==2.0.1
termcolor-whl==1.1.2
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
typing-extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
w3lib==2.0.1
watchdog==2.1.9
webencodings==0.5.1
websockets==10.3
Werkzeug==2.2.2
XlsxWriter==3.0.3
yaspin==2.2.0
zipp==3.8.1
